+++ src/main/java/twilightforest/entity/EntityTFIceSnowball.java
@@ -39,9 +39,9 @@ public class EntityTFIceSnowball extends EntityThrowable {
 	 */
 	public void makeTrail() {
 		for (int i = 0; i < 2; i++) {
-			double dx = posX + 0.5 * (rand.nextDouble() - rand.nextDouble()); 
-			double dy = posY + 0.5 * (rand.nextDouble() - rand.nextDouble()); 
-			double dz = posZ + 0.5 * (rand.nextDouble() - rand.nextDouble()); 
+			double dx = posX + 0.5 * (rand.nextFloat() - rand.nextFloat()); 
+			double dy = posY + 0.5 * (rand.nextFloat() - rand.nextFloat()); 
+			double dz = posZ + 0.5 * (rand.nextFloat() - rand.nextFloat()); 
 			worldObj.spawnParticle("snowballpoof", dx, dy, dz, 0.0D, 0.0D, 0.0D);
 		}
 	}
@@ -86,7 +86,7 @@ public class EntityTFIceSnowball extends EntityThrowable {
 	protected void pop() {
 		for (int i = 0; i < 8; ++i)
 		{
-			this.worldObj.spawnParticle("snowballpoof", this.posX, this.posY, this.posZ, rand.nextGaussian() * 0.05D, rand.nextDouble() * 0.2D, rand.nextGaussian() * 0.05D);
+			this.worldObj.spawnParticle("snowballpoof", this.posX, this.posY, this.posZ, rand.nextGaussian() * 0.05D, rand.nextFloat() * 0.2D, rand.nextGaussian() * 0.05D);
 		}
 		
 		// noise
