+++ src/main/java/twilightforest/entity/EntityTFTwilightWandBolt.java
@@ -36,9 +36,9 @@ public class EntityTFTwilightWandBolt extends EntityThrowable {
 	
 	public void makeTrail() {
 		for (int i = 0; i < 5; i++) {
-			double dx = posX + 0.5 * (rand.nextDouble() - rand.nextDouble()); 
-			double dy = posY + 0.5 * (rand.nextDouble() - rand.nextDouble()); 
-			double dz = posZ + 0.5 * (rand.nextDouble() - rand.nextDouble()); 
+			double dx = posX + 0.5 * (rand.nextFloat() - rand.nextFloat()); 
+			double dy = posY + 0.5 * (rand.nextFloat() - rand.nextFloat()); 
+			double dz = posZ + 0.5 * (rand.nextFloat() - rand.nextFloat()); 
 			
 			double s1 = ((rand.nextFloat() * 0.5F) + 0.5F) * 0.17F;  // color
 			double s2 = ((rand.nextFloat() * 0.5F) + 0.5F) * 0.80F;  // color
@@ -73,7 +73,7 @@ public class EntityTFTwilightWandBolt extends EntityThrowable {
 
         for (int i = 0; i < 8; ++i)
         {
-            this.worldObj.spawnParticle("iconcrack_" + Item.getIdFromItem(Items.ender_pearl), this.posX, this.posY, this.posZ, rand.nextGaussian() * 0.05D, rand.nextDouble() * 0.2D, rand.nextGaussian() * 0.05D);
+            this.worldObj.spawnParticle("iconcrack_" + Item.getIdFromItem(Items.ender_pearl), this.posX, this.posY, this.posZ, rand.nextGaussian() * 0.05D, rand.nextFloat() * 0.2D, rand.nextGaussian() * 0.05D);
         }
 
         if (!this.worldObj.isRemote)
