+++ src/main/java/twilightforest/entity/EntityTFWinterWolf.java
@@ -101,8 +101,8 @@ public class EntityTFWinterWolf extends EntityTFHostileWolf  implements IBreathA
     			double dy = look.yCoord;
     			double dz = look.zCoord;
 
-    			double spread = 5 + this.getRNG().nextDouble() * 2.5;
-    			double velocity = 3.0 + this.getRNG().nextDouble() * 0.15;
+    			double spread = 5 + this.getRNG().nextFloat() * 2.5;
+    			double velocity = 3.0 + this.getRNG().nextFloat() * 0.15;
 
     			// spread flame
     			dx += this.getRNG().nextGaussian() * 0.007499999832361937D * spread;
